requirements.txt
requests>=2.31.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
matplotlib>=3.8.0
pandas>=2.1.0
reportlab>=4.0.0
pillow>=10.0.0
